Module github::activity

Modules

events

Structs

Actor
Actor contains all info on the user generating the event.

EventResponse
EventResponse represents the response for almost all event requests found in activity::events with issues as an exception.

IssueEventResponse
EventResponse represents the response for all issue event requests found in activity::events.

Organisation
Organisation contains all info on the organisation related to the event.

Repository
Documentation References: https://developer.github.com/v3/activity/ All Activity::Events have the same response format. The following structs represent this info found as a json response. Repository contains all info regarding a git repository.

Enums

IssueEventType
IssueEventType is an enumuration of all the different types of Issue Events.

Type Definitions

EventReturnType
EventReturnType is the return type for most event-requests.

IssueEventReturnType
EventReturnType is the return type for issue-event-requests.